% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Calculate the coverage}
\usage{
coverage(true_value, ll, ul, ...)
}
\arguments{
\item{ll}{numeric vector}

\item{ul}{numeric vector}

\item{...}{}

\item{mu_true}{numeric}
}
\value{

}
\description{
Calculate the coverage given a vector of confidence intervals and the true value
}
\examples{
coverage(true_value=0, ll=rnorm(10, -1), ul=rnorm(10, 1))
}
